# openapi.yaml
openapi: 3.0.3
info:
  title: Datamio Audio API
  description: |
    FastAPI endpoint for uploading multi-domain audio datasets to Hugging Face with audio splitting capabilities.

    ## Authentication

    All API endpoints (except `GET /`) require an API key passed in the `X-API-Key` header.

    ### Key Types

    | Key Type | Source | Capabilities |
    |----------|--------|--------------|
    | Admin Key | `ADMIN_API_KEY` environment variable | Full access + can create/manage API keys |
    | User Key | Generated via `POST /api/keys` | Access all endpoints except key creation |

    ### Usage

    Include the API key in your request headers:
    ```
    X-API-Key: your-api-key-here
    ```
  version: 1.0.0
servers:
  - url: http://localhost:8001
    description: Local development server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Set ADMIN_API_KEY env var for admin access.
  schemas:
...
    Segment:
      type: object
      properties:
        start:
          type: number
          format: float
          description: Start time in seconds
        end:
          type: number
          format: float
          description: End time in seconds
        url:
          type: string
          description: URL to the audio chunk
    SplitsResponse:
      type: object
      properties:
        segments:
          type: array
          items:
            $ref: '#/components/schemas/Segment'
        count:
          type: integer
          description: Number of segments
...
    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: object
              properties:
                error:
                  type: string
                details:
                  type: string
                stack:
                  type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
security:
  - ApiKeyAuth: []
paths:
...
  /api/splits/file:
    post:
      tags:
        - Splits (Sync)
      summary: Split Single Audio File
      description: Split audio file by silence detection. Returns segments with chunk file paths.
      operationId: splitFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplitsRequest'
      responses:
        '200':
          description: Audio successfully split
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SplitsResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/splits/batch:
    post:
      tags:
        - Splits (Sync)
      summary: Split Multiple Audio Files
      description: Split multiple audio files by silence detection. Returns segments for each file.
      operationId: splitBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSplitsRequest'
      responses:
        '200':
          description: Batch successfully processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSplitsResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/splits/file/job:
    post:
      tags:
        - Splits (Async)
      summary: Create Split File Job
      description: Create an async job for splitting audio file. Returns immediately with job ID.
      operationId: createSplitFileJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplitsRequest'
      responses:
        '200':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SplitJobResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Job creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/splits/batch/job:
    post:
      tags:
        - Splits (Async)
      summary: Create Batch Split Job
      description: Create an async job for splitting multiple audio files. Returns immediately with job ID.
      operationId: createSplitBatchJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSplitsRequest'
      responses:
        '200':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SplitJobResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Job creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/splits/job/{job_id}:
    get:
      tags:
        - Splits (Async)
      summary: Get Split Job Status
      description: Get status of a specific split job
      operationId: getSplitJobStatus
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: The job ID
      responses:
        '200':
          description: Job status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SplitJobResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Splits (Async)
      summary: Delete Split Job
      description: Delete a split job from the database
      operationId: deleteSplitJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: The job ID to delete
      responses:
        '200':
          description: Job deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/splits/jobs:
    get:
      tags:
        - Splits (Async)
      summary: Get All Split Jobs
      description: Get all split jobs for a user. **Requires admin API key.**
      operationId: getAllSplitJobs
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
            default: default
          description: User ID to filter jobs
      responses:
        '200':
          description: List of split jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SplitJobResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only admin can list all jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/upload-audio-dataset:
    post:
      tags:
        - Hugging Face
      summary: Upload Audio Dataset
      description: Create an upload job for audio dataset to Hugging Face. Returns immediately with job ID.
      operationId: uploadAudioDataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadAudioDatasetRequest'
      responses:
        '200':
          description: Upload job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Job creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/jobs:
    get:
      tags:
        - Hugging Face
      summary: Get All Upload Jobs
      description: Get all upload jobs for a user. **Requires admin API key.**
      operationId: getAllJobs
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
            default: default
          description: User ID to filter jobs
      responses:
        '200':
          description: List of upload jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only admin can list all jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/keys/{key_id}/revoke:
    post:
      tags:
        - API Key Management
      summary: Revoke API Key
      description: Revoke (disable) an API key without permanently deleting it. **Requires admin API key.**
      operationId: revokeApiKey
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
          description: The API key ID to revoke
      responses:
        '200':
          description: API key revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only admin can revoke API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'